% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canaryNGOs-package.R
\docType{package}
\name{canaryNGOs-package}
\alias{canaryNGOs-package}
\alias{canaryNGOs}
\title{The 'canaryNGOs' package.}
\description{
Research on how restrictions on NGOs foreshadow other forms of
  human rights repression
}
