% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{pts}
\alias{pts}
\title{Convert mms to pts}
\usage{
pts(x)
}
\arguments{
\item{x}{a numeric value (in millimeters)}
}
\value{
A numeric value (in points)
}
\description{
Convert units specified in millimeters to typographic points. This is especially helpful when working with ggplot geoms that use size parameters
}
\examples{
library(ggplot2)

ggplot(mtcars, aes(x = mpg, y = wt)) +
  geom_point() +
  annotate(geom = "text", x = 20, y = 4,
           label = "Here's a label", size = pts(11))
}
