% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{theme_ngo}
\alias{theme_ngo}
\title{theme_ngo}
\usage{
theme_ngo(base_size = 9, base_family = "Inter Light")
}
\arguments{
\item{base_size}{base font size (default is 9)}

\item{base_family}{base font family (default is Roboto Condensed)}
}
\description{
A custom ggplot2 theme used throughout this project
}
\examples{
library(ggplot2)

ggplot(mtcars, aes(x = mpg, y = wt)) +
  geom_point() +
  theme_ngo()
}
